compaction mode
