+++ indexing-service/src/main/java/org/apache/druid/indexing/compact/OverlordCompactionScheduler.java
@@ -499,7 +499,7 @@ public class OverlordCompactionScheduler implements CompactionScheduler
           updateRequest.getEngine()
       );
     } else {
-      return new CompactionSimulateResult(Collections.emptyMap());
+      return new CompactionSimulateResult(Collections.emptyMap(), null);
     }
   }
