+++ server/src/test/java/org/apache/druid/server/compaction/MostFragmentedIntervalFirstPolicyTest.java
@@ -22,8 +22,8 @@ package org.apache.druid.server.compaction;
 import nl.jqno.equalsverifier.EqualsVerifier;
 import org.apache.druid.jackson.DefaultObjectMapper;
 import org.apache.druid.java.util.common.HumanReadableBytes;
-import org.apache.druid.segment.TestDataSource;
-import org.apache.druid.server.coordinator.CreateDataSegments;
+import org.apache.druid.java.util.common.Intervals;
+import org.apache.druid.segment.TestSegmentUtils;
 import org.apache.druid.timeline.DataSegment;
 import org.junit.Test;
 import org.junit.jupiter.api.Assertions;
@@ -34,9 +34,11 @@ import java.util.List;
 public class MostFragmentedIntervalFirstPolicyTest
 {
   private static final DataSegment SEGMENT =
-      CreateDataSegments.ofDatasource(TestDataSource.WIKI).eachOfSizeInMb(100).get(0);
+      TestSegmentUtils.makeSegment("foo", "1", Intervals.ETERNITY);
+  private static final DataSegment SEGMENT2 =
+      TestSegmentUtils.makeSegment("foo", "2", Intervals.ETERNITY);
   private static final CompactionCandidate.ProposedCompaction PROPOSED_COMPACTION =
-      CompactionCandidate.ProposedCompaction.from(List.of(SEGMENT), null);
+      CompactionCandidate.ProposedCompaction.from(List.of(SEGMENT, SEGMENT2), null);
 
   private static final CompactionStatistics DUMMY_COMPACTION_STATS = CompactionStatistics.create(1L, 1L, 1L);
 
@@ -52,7 +54,7 @@ public class MostFragmentedIntervalFirstPolicyTest
   }
 
   @Test
-  public void test_checkEligibilityForCompaction_fails_ifUncompactedCountLessThanCutoff()
+  public void test_createCandidate_fails_ifUncompactedCountLessThanCutoff()
   {
     final int minUncompactedCount = 10_000;
     final MostFragmentedIntervalFirstPolicy policy = new MostFragmentedIntervalFirstPolicy(
@@ -63,20 +65,23 @@ public class MostFragmentedIntervalFirstPolicyTest
         null
     );
 
-    final CompactionEligibility eligibility1 =
+    final CompactionStatus eligibility1 =
         eligibilityBuilder().compacted(DUMMY_COMPACTION_STATS).uncompacted(createStats(1, 100L)).build();
+    final CompactionCandidate candidate1 = policy.createCandidate(PROPOSED_COMPACTION, eligibility1);
     Assertions.assertEquals(
-        CompactionEligibility.fail("Uncompacted segments[1] in interval must be at least [10,000]"),
-        policy.checkEligibilityForCompaction(PROPOSED_COMPACTION, eligibility1)
+        "Uncompacted segments[1] in interval must be at least [10,000]",
+        candidate1.getPolicyNote()
     );
+    Assertions.assertEquals(CompactionMode.NOT_APPLICABLE, candidate1.getMode());
 
-    final CompactionEligibility eligibility2 =
+    final CompactionStatus eligibility2 =
         eligibilityBuilder().compacted(DUMMY_COMPACTION_STATS).uncompacted(createStats(10_001, 100L)).build();
-    Assertions.assertEquals(eligibility2, policy.checkEligibilityForCompaction(PROPOSED_COMPACTION, eligibility2));
+    final CompactionCandidate candidate2 = policy.createCandidate(PROPOSED_COMPACTION, eligibility2);
+    Assertions.assertEquals(CompactionMode.FULL_COMPACTION, candidate2.getMode());
   }
 
   @Test
-  public void test_checkEligibilityForCompaction_fails_ifUncompactedBytesLessThanCutoff()
+  public void test_createCandidate_fails_ifUncompactedBytesLessThanCutoff()
   {
     final HumanReadableBytes minUncompactedBytes = HumanReadableBytes.valueOf(10_000);
     final MostFragmentedIntervalFirstPolicy policy = new MostFragmentedIntervalFirstPolicy(
@@ -87,20 +92,20 @@ public class MostFragmentedIntervalFirstPolicyTest
         null
     );
 
-    final CompactionEligibility eligibility1 =
+    final CompactionStatus eligibility1 =
         eligibilityBuilder().compacted(DUMMY_COMPACTION_STATS).uncompacted(createStats(1, 100L)).build();
-    Assertions.assertEquals(
-        CompactionEligibility.fail("Uncompacted bytes[100] in interval must be at least [10,000]"),
-        policy.checkEligibilityForCompaction(PROPOSED_COMPACTION, eligibility1)
-    );
+    final CompactionCandidate candidate1 = policy.createCandidate(PROPOSED_COMPACTION, eligibility1);
+    Assertions.assertEquals("Uncompacted bytes[100] in interval must be at least [10,000]", candidate1.getPolicyNote());
+    Assertions.assertEquals(CompactionMode.NOT_APPLICABLE, candidate1.getMode());
 
-    final CompactionEligibility eligibility2 =
+    final CompactionStatus eligibility2 =
         eligibilityBuilder().compacted(DUMMY_COMPACTION_STATS).uncompacted(createStats(100, 10_000L)).build();
-    Assertions.assertEquals(eligibility2, policy.checkEligibilityForCompaction(PROPOSED_COMPACTION, eligibility2));
+    final CompactionCandidate candidate2 = policy.createCandidate(PROPOSED_COMPACTION, eligibility2);
+    Assertions.assertEquals(CompactionMode.FULL_COMPACTION, candidate2.getMode());
   }
 
   @Test
-  public void test_checkEligibilityForCompaction_fails_ifAvgSegmentSizeGreaterThanCutoff()
+  public void test_createCandidate_fails_ifAvgSegmentSizeGreaterThanCutoff()
   {
     final HumanReadableBytes maxAvgSegmentSize = HumanReadableBytes.valueOf(100);
     final MostFragmentedIntervalFirstPolicy policy = new MostFragmentedIntervalFirstPolicy(
@@ -111,15 +116,18 @@ public class MostFragmentedIntervalFirstPolicyTest
         null
     );
 
-    final CompactionEligibility eligibility1 =
+    final CompactionStatus eligibility1 =
         eligibilityBuilder().compacted(DUMMY_COMPACTION_STATS).uncompacted(createStats(1, 10_000L)).build();
+    final CompactionCandidate candidate1 = policy.createCandidate(PROPOSED_COMPACTION, eligibility1);
     Assertions.assertEquals(
-        CompactionEligibility.fail("Average size[10,000] of uncompacted segments in interval must be at most [100]"),
-        policy.checkEligibilityForCompaction(PROPOSED_COMPACTION, eligibility1)
+        "Average size[10,000] of uncompacted segments in interval must be at most [100]",
+        candidate1.getPolicyNote()
     );
-    final CompactionEligibility eligibility2 =
+    Assertions.assertEquals(CompactionMode.NOT_APPLICABLE, candidate1.getMode());
+    final CompactionStatus eligibility2 =
         eligibilityBuilder().compacted(DUMMY_COMPACTION_STATS).uncompacted(createStats(1, 100L)).build();
-    Assertions.assertEquals(eligibility2, policy.checkEligibilityForCompaction(PROPOSED_COMPACTION, eligibility2));
+    final CompactionCandidate candidate2 = policy.createCandidate(PROPOSED_COMPACTION, eligibility2);
+    Assertions.assertEquals(CompactionMode.FULL_COMPACTION, candidate2.getMode());
   }
 
   @Test
@@ -133,16 +141,13 @@ public class MostFragmentedIntervalFirstPolicyTest
         null
     );
 
-    final CompactionEligibility eligibility1 =
+    final CompactionStatus eligibility1 =
         eligibilityBuilder().compacted(DUMMY_COMPACTION_STATS).uncompacted(createStats(1, 1_000L)).build();
-    final CompactionEligibility eligibility2 =
+    final CompactionStatus eligibility2 =
         eligibilityBuilder().compacted(DUMMY_COMPACTION_STATS).uncompacted(createStats(2, 500L)).build();
 
-    Assertions.assertEquals(eligibility1, policy.checkEligibilityForCompaction(PROPOSED_COMPACTION, eligibility1));
-    Assertions.assertEquals(eligibility2, policy.checkEligibilityForCompaction(PROPOSED_COMPACTION, eligibility2));
-
-    final CompactionCandidate candidateA = eligibility1.createCandidate(PROPOSED_COMPACTION);
-    final CompactionCandidate candidateB = eligibility2.createCandidate(PROPOSED_COMPACTION);
+    final CompactionCandidate candidateA = policy.createCandidate(PROPOSED_COMPACTION, eligibility1);
+    final CompactionCandidate candidateB = policy.createCandidate(PROPOSED_COMPACTION, eligibility2);
     Assertions.assertTrue(policy.compareCandidates(candidateA, candidateB) > 0);
     Assertions.assertTrue(policy.compareCandidates(candidateB, candidateA) < 0);
   }
@@ -158,17 +163,13 @@ public class MostFragmentedIntervalFirstPolicyTest
         null
     );
 
-    final CompactionEligibility eligibility1 =
+    final CompactionStatus eligibility1 =
         eligibilityBuilder().compacted(DUMMY_COMPACTION_STATS).uncompacted(createStats(1, 1000L)).build();
-    final CompactionEligibility eligibility2 =
+    final CompactionStatus eligibility2 =
         eligibilityBuilder().compacted(DUMMY_COMPACTION_STATS).uncompacted(createStats(2, 1000L)).build();
 
-
-    Assertions.assertEquals(eligibility1, policy.checkEligibilityForCompaction(PROPOSED_COMPACTION, eligibility1));
-    Assertions.assertEquals(eligibility2, policy.checkEligibilityForCompaction(PROPOSED_COMPACTION, eligibility2));
-
-    final CompactionCandidate candidateA = eligibility1.createCandidate(PROPOSED_COMPACTION);
-    final CompactionCandidate candidateB = eligibility2.createCandidate(PROPOSED_COMPACTION);
+    final CompactionCandidate candidateA = policy.createCandidate(PROPOSED_COMPACTION, eligibility1);
+    final CompactionCandidate candidateB = policy.createCandidate(PROPOSED_COMPACTION, eligibility2);
     Assertions.assertTrue(policy.compareCandidates(candidateA, candidateB) > 0);
     Assertions.assertTrue(policy.compareCandidates(candidateB, candidateA) < 0);
   }
@@ -184,16 +185,13 @@ public class MostFragmentedIntervalFirstPolicyTest
         null
     );
 
-    final CompactionEligibility eligibility1 =
+    final CompactionStatus eligibility1 =
         eligibilityBuilder().compacted(DUMMY_COMPACTION_STATS).uncompacted(createStats(10, 500L)).build();
-    final CompactionEligibility eligibility2 =
+    final CompactionStatus eligibility2 =
         eligibilityBuilder().compacted(DUMMY_COMPACTION_STATS).uncompacted(createStats(10, 1000L)).build();
 
-    Assertions.assertEquals(eligibility1, policy.checkEligibilityForCompaction(PROPOSED_COMPACTION, eligibility1));
-    Assertions.assertEquals(eligibility2, policy.checkEligibilityForCompaction(PROPOSED_COMPACTION, eligibility2));
-
-    final CompactionCandidate candidateA = eligibility1.createCandidate(PROPOSED_COMPACTION);
-    final CompactionCandidate candidateB = eligibility2.createCandidate(PROPOSED_COMPACTION);
+    final CompactionCandidate candidateA = policy.createCandidate(PROPOSED_COMPACTION, eligibility1);
+    final CompactionCandidate candidateB = policy.createCandidate(PROPOSED_COMPACTION, eligibility2);
     Assertions.assertTrue(policy.compareCandidates(candidateA, candidateB) < 0);
     Assertions.assertTrue(policy.compareCandidates(candidateB, candidateA) > 0);
   }
@@ -209,16 +207,13 @@ public class MostFragmentedIntervalFirstPolicyTest
         null
     );
 
-    final CompactionEligibility eligibility1 =
+    final CompactionStatus eligibility1 =
         eligibilityBuilder().compacted(DUMMY_COMPACTION_STATS).uncompacted(createStats(100, 25)).build();
-    final CompactionEligibility eligibility2 =
+    final CompactionStatus eligibility2 =
         eligibilityBuilder().compacted(DUMMY_COMPACTION_STATS).uncompacted(createStats(400, 100)).build();
 
-    Assertions.assertEquals(eligibility1, policy.checkEligibilityForCompaction(PROPOSED_COMPACTION, eligibility1));
-    Assertions.assertEquals(eligibility2, policy.checkEligibilityForCompaction(PROPOSED_COMPACTION, eligibility2));
-
-    final CompactionCandidate candidateA = eligibility1.createCandidate(PROPOSED_COMPACTION);
-    final CompactionCandidate candidateB = eligibility2.createCandidate(PROPOSED_COMPACTION);
+    final CompactionCandidate candidateA = policy.createCandidate(PROPOSED_COMPACTION, eligibility1);
+    final CompactionCandidate candidateB = policy.createCandidate(PROPOSED_COMPACTION, eligibility2);
     Assertions.assertEquals(0, policy.compareCandidates(candidateA, candidateB));
     Assertions.assertEquals(0, policy.compareCandidates(candidateB, candidateA));
   }
@@ -260,7 +255,7 @@ public class MostFragmentedIntervalFirstPolicyTest
   }
 
   @Test
-  public void test_checkEligibilityForCompaction_returnsIncrementalCompaction_whenRatioBelowThreshold()
+  public void test_createCandidate_returnsIncrementalCompaction_whenRatioBelowThreshold()
   {
     // Set threshold to 0.5 (50%)
     final MostFragmentedIntervalFirstPolicy policy = new MostFragmentedIntervalFirstPolicy(
@@ -273,27 +268,23 @@ public class MostFragmentedIntervalFirstPolicyTest
 
     final CompactionStatistics compacted = CompactionStatistics.create(1200L, 10, 1L);
     final CompactionStatistics uncompacted = CompactionStatistics.create(400L, 100, 1L);
-    final CompactionEligibility candidate = eligibilityBuilder()
+    final CompactionStatus eligibility = eligibilityBuilder()
         .compacted(compacted)
         .uncompacted(uncompacted)
         .uncompactedSegments(List.of(SEGMENT))
         .build();
 
+    final CompactionCandidate candidate = policy.createCandidate(PROPOSED_COMPACTION, eligibility);
+    Assertions.assertEquals("Uncompacted bytes ratio[0.25] is below threshold[0.50]", candidate.getPolicyNote());
+    Assertions.assertEquals(CompactionMode.INCREMENTAL_COMPACTION, candidate.getMode());
     Assertions.assertEquals(
-        CompactionEligibility.builder(
-                                 CompactionEligibility.State.INCREMENTAL_COMPACTION,
-                                 "Uncompacted bytes ratio[0.25] is below threshold[0.50]"
-                             )
-                             .compacted(compacted)
-                             .uncompacted(uncompacted)
-                             .uncompactedSegments(List.of(SEGMENT))
-                             .build(),
-        policy.checkEligibilityForCompaction(PROPOSED_COMPACTION, candidate)
+        CompactionCandidate.ProposedCompaction.from(List.of(SEGMENT), null),
+        candidate.getProposedCompaction()
     );
   }
 
   @Test
-  public void test_checkEligibilityForCompaction_returnsFullCompaction_whenRatioAboveThreshold()
+  public void test_createCandidate_returnsFullCompaction_whenRatioAboveThreshold()
   {
     // Set threshold to 0.5 (50%)
     final MostFragmentedIntervalFirstPolicy policy = new MostFragmentedIntervalFirstPolicy(
@@ -304,20 +295,18 @@ public class MostFragmentedIntervalFirstPolicyTest
         null
     );
 
-    final CompactionEligibility eligibility =
+    final CompactionStatus eligibility =
         eligibilityBuilder()
             .compacted(CompactionStatistics.create(500L, 5, 1))
             .uncompacted(CompactionStatistics.create(600L, 100, 1))
             .build();
+    final CompactionCandidate candidate = policy.createCandidate(PROPOSED_COMPACTION, eligibility);
 
-    Assertions.assertEquals(
-        eligibility,
-        policy.checkEligibilityForCompaction(PROPOSED_COMPACTION, eligibility)
-    );
+    Assertions.assertEquals(CompactionMode.FULL_COMPACTION, candidate.getMode());
   }
 
   @Test
-  public void test_checkEligibilityForCompaction_returnsFullCompaction_whenThresholdIsDefault()
+  public void test_createCandidate_returnsFullCompaction_whenThresholdIsDefault()
   {
     // Default threshold is 0.0
     final MostFragmentedIntervalFirstPolicy policy = new MostFragmentedIntervalFirstPolicy(
@@ -329,13 +318,14 @@ public class MostFragmentedIntervalFirstPolicyTest
     );
 
     // With default threshold 0.0, any positive ratio >= 0.0, so always FULL_COMPACTION_ELIGIBLE
-    final CompactionEligibility eligibility =
+    final CompactionStatus eligibility =
         eligibilityBuilder()
             .compacted(CompactionStatistics.create(1_000L, 10, 1))
             .uncompacted(CompactionStatistics.create(100L, 100, 1))
             .build();
+    final CompactionCandidate candidate = policy.createCandidate(PROPOSED_COMPACTION, eligibility);
 
-    Assertions.assertEquals(eligibility, policy.checkEligibilityForCompaction(PROPOSED_COMPACTION, eligibility));
+    Assertions.assertEquals(CompactionMode.FULL_COMPACTION, candidate.getMode());
   }
 
   private CompactionStatistics createStats(int numSegments, long avgSizeBytes)
@@ -343,9 +333,9 @@ public class MostFragmentedIntervalFirstPolicyTest
     return CompactionStatistics.create(avgSizeBytes * numSegments, numSegments, 1L);
   }
 
-  private static CompactionEligibility.CompactionEligibilityBuilder eligibilityBuilder()
+  private static CompactionStatus.CompactionStatusBuilder eligibilityBuilder()
   {
-    return CompactionEligibility.builder(CompactionEligibility.State.FULL_COMPACTION, "approve")
-                                .uncompactedSegments(List.of());
+    return CompactionStatus.builder(CompactionStatus.State.ELIGIBLE, "approve")
+                           .uncompactedSegments(List.of());
   }
 }
